file I/O - 0

+++ file_io/0-read_textfile.c
@@ -1,4 +1,5 @@
 #include "main.h"
+
 /**
 * read_textfile - reads a text file and prints it to the POSIX standard output.
 * @filename: pointer to the name of the file to be read.
@@ -11,7 +12,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 {
 if (filename == NULL)
 	return (0);
-int fd = open(filename, O_RDONLY);
+ssize_t fd = open(filename, O_RDONLY);
 
 	if (fd == -1)
 		return (0);
